Create opcode info lib, create next and step
Basically, what was disassemble.cpp before is now opcodeinfo,
a small library for extracting information about opcodes.
This shall be used extensively in the debugger (hopefully)
Introduced a few new functions in the CPU, mainly a nextaddr() function
that finds the next address the pc will be at after running a given
instruction.
Finished step() and next() debugger commands.