Fixed macro bug.
The 'limits.h' header file in POSIX systems #defines a macro,
'LINE_MAX', that would conflict with the name of a constant. The
constant was thus renamed.
The constant didn't really appear anywhere where it was supposed to
appear, so I decided to search for the magic numbers that the constant
defined and replaced them.