Split PPU output buffer and texture info.
The PPU now owns its output buffer, which is a simple Array2D. Thanks
also to the coordinate flipping change in the last commit, the
drawpixel() function is reduced to a single store.
A lot of files had to be changed in order not to let everything break.
In particular, Video::Canvas doesn't exist anymore, and both Emulator
and PPU have getters.