Modify flags struct to occupy less space.
This was a pet peeve of mine. You can have the CPU flags only take the
size of a byte as long you use the BitField type and do proper
assignments (instead of implicit casts to bool, for which I was bitten).
I decided to implement this space save in this commit.