Fix inaccuracy bug in Donkey Kong.
There was a bug in Donkey Kong where the score would get garbage values,
and where mario's jump was noticeably bigger. This was caused by an
inaccurate SBC instruction: the result was calculated using this
formula:

        res = acc + ~val + carry

however, to work, the middle operand must be of an unsigned type (I'm
not sure why, code generation for some reason trips on this).
The current code now offloads the work to instr_adc, which means the
middle operand is now back to an unsigned type, and also that the
overflow flag is calculated with a flipped "val" value (this was
apparently an oversight I had made when first writing the function).