Change MMC1::read_rom so that it has no branches.
I felt that optimizing this one function was very worth it since it gets
called so much (if the game uses mapper 1, that is). The new MMC1 class
uses (an array of) 2 pointers for the first and second bank, that to
the right bank according to the PRG mode. Consequently we do not need to
save to PRG mode anymore, but we do have to keep the first and last bank
somewhere, so that the pointers can point to something.