Fix bugs in mirroring and NROM.
All mirrorings except for horizontal were bugged due to me forgetting
how to use bitmask() and making an off by one error, oops!
Also, NROM::read_rom() incorrectly calculated the offset assuming the
ROM was always 0x8000 in size (this is wrong; NROM allows for a size of
0x4000 and mirrors the rest of the address space): this was also fixed.

Interestingly, the formula I always used before implementing mappers
was also incorrect, but could still, somehow, calculate the right
address. The formula was this:

    u16 offset = addr - 0x8000;
    u16 start = prgrom.size() - 0x8000;
    u16 eff_addr = start + offset;
    return prgrom[eff_addr];

And it contains two overflows when prgrom.size() is 0x4000.