Fix changes to mirroring exhausting bus.
Bus exhaustion can be reached easily in tetris by switching back and
forth between the title screen and the level select screen, so I decided
to implement a fix.
There are two implementation, one commented out. I haven't checked yet
which one is the better one.

I've also fixed a minor bug with the select and start keys being swapped
in the controller.